backend/requirements.txt
alembic==1.16.4
aniso8601==10.0.1
blinker==1.9.0
captcha==0.7.1
charset-normalizer==3.4.3
click==8.2.1
et_xmlfile==2.0.0
Flask==3.1.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
flask-marshmallow==1.3.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.1.1
greenlet==3.2.4
itsdangerous==2.2.0
Jinja2==3.1.6
langid==1.1.6
Mako==1.3.10
MarkupSafe==3.0.2
marshmallow==4.0.0
marshmallow-sqlalchemy==1.4.2
mysqlclient==2.2.7
numpy==2.3.2
openpyxl==3.1.5
pandas==2.3.1
pillow==11.3.0
praat-parselmouth==0.4.6
praat-textgrids==1.4.0
praatio==6.2.0
psutil==7.0.0
PyJWT==2.10.1
pympi-ling==1.70.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
setuptools==80.9.0
six==1.17.0
SQLAlchemy==2.0.42
TextGrid==1.6.1
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
